requirements.txt
# Core Framework
fastmcp>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async Support
asyncio>=3.4.3
aiohttp>=3.9.0
aiofiles>=23.0.0

# Database & Cache
redis>=5.0.0
asyncpg>=0.29.0
sqlalchemy>=2.0.0

# Volatility3 (installed separately in Docker)
# volatility3>=2.5.0

# HTTP & API
httpx>=0.27.0
requests>=2.31.0

# Data Processing
pyyaml>=6.0.0
orjson>=3.9.0

# Security
python-jose>=3.3.0
passlib>=1.7.4

# Logging & Monitoring
structlog>=24.0.0
python-json-logger>=2.0.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
tenacity>=8.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Type Checking
mypy>=1.8.0
types-redis>=4.6.0
types-pyyaml>=6.0.0
types-requests>=2.31.0

# Development
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
pre-commit>=3.6.0
python-dateutil>=2.8.0